+++ codecs/vp8_packet.go
@@ -8,7 +8,7 @@ const (
 )
 
 // Payload fragments a VP8 packet across one or more byte arrays
-func (p *VP8Payloader) Payload(mtu int, payload []byte) [][]byte {
+func (p *VP8Payloader) Payload(mtu uint16, payload []byte) [][]byte {
 	/*
 	 * https://tools.ietf.org/html/rfc7741#section-4.2
 	 *
@@ -30,7 +30,7 @@ func (p *VP8Payloader) Payload(mtu int, payload []byte) [][]byte {
 	 *     first packet of each encoded frame.
 	 */
 
-	maxFragmentSize := mtu - vp8HeaderSize
+	maxFragmentSize := int(mtu) - vp8HeaderSize
 
 	payloadData := payload
 	payloadDataRemaining := len(payload)
